+++ tests/test_delete_signal.py
@@ -111,3 +111,15 @@ class TestDeleteSignal(TransactionTestCase):
         mocked_signal.assert_called_once()
         signal_called_with_first = mocked_signal.call_args_list[0].kwargs
         self.assertEqual(signal_called_with_first["sender"], Post)
+
+    @patch("bulk_tracker.signals.post_delete_signal.send")
+    @patch("bulk_tracker.signals.post_delete_signal.send_robust")
+    def test_should_use_robust_send_if_is_robust_is_true_in_tracking_info(self, mocked_signal_robust, mocked_signal):
+        # Arrange
+
+        # Act
+        self.author_john.delete(tracking_info_=TrackingInfo(is_robust=True))
+
+        # Assert
+        mocked_signal.assert_not_called()
+        mocked_signal_robust.assert_called_once()