+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 This was causing the signals to be sent before the object was actually created/updated/deleted, Causing a race condition.
 This has been fixed by using `transaction.on_commit()`. 
 This will ensure that the signals are sent after the transaction is committed.
+- Add support for send_robust by using `tracking_info_TrackingInfo(is_robust=True)`. `is_robust=False` by default.
 
 ## 0.0.7 (2023-07-06)
 - A fix where `send_post_create_signal()` was being called twice when creating an object through `BulkTrackerQuerySet.create()`